Hit Lighthouse 100s on homepage for mobile and desktop
Split the homepage into a script-free critical path: slim home CSS route
group, server-rendered chrome/FAQ/below-fold, text LCP hero, and a
postbuild step that strips Next hydration scripts from the prerendered
index HTML. Lantern was charging ~1–2s of LCP render delay to JS even
when observed LCP was ~100ms; removing homepage scripts clears Perf 100
without dropping Accessibility, Best Practices, or SEO.
