+++ app/(home)/layout.tsx
@@ -1,0 +1,16 @@
+import { SiteHeader } from "@/components/site-header"
+import { StaticSiteFooter } from "@/components/static-site-footer"
+
+import "./home.css"
+
+export default function HomeLayout({
+	children,
+}: Readonly<{ children: React.ReactNode }>) {
+	return (
+		<>
+			<SiteHeader />
+			{children}
+			<StaticSiteFooter />
+		</>
+	)
+}
