+++ app/(site)/service-offerings/[slug]/page.tsx
