+++ app/layout.tsx
@@ -1,40 +1,11 @@
 import type { Metadata, Viewport } from "next"
-import { Archivo, Manrope } from "next/font/google"
 import { cacheLife, cacheTag } from "next/cache"
 import { Suspense } from "react"
 
-import { AnalyticsLoader } from "@/components/analytics-loader"
-import { CommandMenuLoader } from "@/components/command-menu-loader"
 import { JsonLd } from "@/components/json-ld"
-import { SiteFooter } from "@/components/site-footer"
-import { SiteHeader } from "@/components/site-header"
 import { getCollection } from "@/lib/content"
 import { localBusinessJsonLd, websiteJsonLd } from "@/lib/seo"
 import { siteConfig } from "@/lib/site"
-
-import "./globals.css"
-
-const manrope = Manrope({
-	subsets: ["latin"],
-	display: "swap",
-	variable: "--font-manrope",
-	preload: true,
-	weight: ["400", "700"],
-})
-
-/*
-  Display face. A squared grotesque against Manrope's humanist body - the
-  contrast is what gives headings their own voice instead of running the body
-  face at weight 800. Only the two display weights are loaded.
-*/
-const archivo = Archivo({
-	subsets: ["latin"],
-	display: "swap",
-	variable: "--font-archivo",
-	/* Body font covers first paint; Archivo is for display headings. */
-	preload: false,
-	weight: ["700", "800"],
-})
 
 export const metadata: Metadata = {
 	metadataBase: new URL(siteConfig.url),
@@ -108,25 +79,18 @@ export default function RootLayout({
 	children,
 }: Readonly<{ children: React.ReactNode }>) {
 	return (
-		<html
-			lang="en"
-			className={`${manrope.variable} ${archivo.variable}`}
-		>
-			<body className={manrope.className}>
+		<html lang="en">
+			<body>
 				<a
 					className="sr-only z-[100] rounded-br-lg bg-white p-3 text-black focus:not-sr-only focus:fixed focus:top-0 focus:left-0"
 					href="#main-content"
 				>
 					Skip to content
 				</a>
-				<SiteHeader />
 				{children}
-				<SiteFooter />
-				<CommandMenuLoader />
 				<Suspense fallback={null}>
 					<RootJsonLd />
 				</Suspense>
-				<AnalyticsLoader />
 			</body>
 		</html>
 	)
