+++ lib/home-content.ts
@@ -1,0 +1,129 @@
+export const serviceGroups = [
+	{
+		title: "Residential Plumbing",
+		image: "/images/work/precision-valve-installation.webp",
+		description:
+			"Repairs and installations completed cleanly, honestly, and to code.",
+		items: [
+			"Leak detection & repair",
+			"Fixture installation & repair",
+			"Pipe replacement & repiping",
+			"Drain cleaning & unclogging",
+		],
+	},
+	{
+		title: "Commercial Plumbing",
+		image: "/images/work/commercial-plumbing-installation.webp",
+		description:
+			"Dependable commercial service that keeps your property operating.",
+		items: [
+			"Commercial water heaters",
+			"Backflow prevention",
+			"Code compliance updates",
+			"Restaurant & kitchen plumbing",
+		],
+	},
+	{
+		title: "New Construction",
+		image: "/images/work/new-construction-rough-in.webp",
+		description:
+			"Code-compliant plumbing planned from rough-in through final testing.",
+		items: [
+			"Plumbing system design",
+			"Rough-in plumbing",
+			"Fixture installation",
+			"Final inspections & testing",
+		],
+	},
+]
+
+export const workGallery = [
+	{
+		src: "/images/work/engineered-retaining-wall.webp",
+		alt: "Engineered septic retaining wall and control panel in Santa Cruz County",
+		caption: "Engineered retaining wall and control panel",
+	},
+	{
+		src: "/images/work/multi-tank-excavation.webp",
+		alt: "Multi-tank septic excavation underway on a hillside property",
+		caption: "Multi-tank excavation in progress",
+	},
+	{
+		src: "/images/work/completed-multi-tank.webp",
+		alt: "Completed multi-tank engineered septic installation",
+		caption: "Completed multi-tank installation",
+	},
+	{
+		src: "/images/work/advanced-septic-control-panel.webp",
+		alt: "Advanced septic control panel and system access points",
+		caption: "Advanced septic control system",
+	},
+] as const
+
+export const trustedBrands = [
+	{ src: "/images/partners/bradfordwhite.webp", alt: "Bradford White" },
+	{ src: "/images/partners/navien.webp", alt: "Navien" },
+	{ src: "/images/partners/kohler.webp", alt: "Kohler" },
+	{ src: "/images/partners/grohe.webp", alt: "Grohe" },
+	{ src: "/images/partners/grundfos.webp", alt: "Grundfos" },
+] as const
+
+export const heroStats = [
+	{ label: "Licensed", value: "CA CSLB #1087260" },
+	{ label: "Local rating", value: "4.9 out of 5" },
+	{ label: "Office hours", value: "Mon to Fri, 9 to 5" },
+] as const
+
+export const trustPoints = [
+	"Family owned & operated",
+	"Licensed & insured",
+	"Satisfaction guaranteed",
+	"4.9 local rating",
+] as const
+
+export const reasons = [
+	["No-pressure advice", "Only the work your property needs."],
+	["Transparent pricing", "Clear pricing before work begins."],
+	["Clean workmanship", "Respectful service and a clean finish."],
+	["Long-term reliability", "Repairs and systems designed to last."],
+] as const
+
+export const testimonials = [
+	[
+		"As a real estate professional, I love working with Wade’s Plumbing. They are courteous, professional, and always make sure my clients are taken care of.",
+		"Bailey Cotrona",
+	],
+	[
+		"They know their stuff, do great work, and are a pleasure to work with. They made me feel comfortable with a terrible situation that shut down my house.",
+		"Aaron Berger",
+	],
+] as const
+
+export const faqs = [
+	{
+		question: "How often should I have my septic tank pumped?",
+		answer:
+			"Most septic tanks should be pumped every 3 to 5 years, depending on household size, wastewater volume, and tank size. We recommend a professional inspection to determine the right schedule for your system.",
+	},
+	{
+		question: "What are the signs of a failing septic system?",
+		answer:
+			"Warning signs include slow drains, gurgling sounds, sewage backups, soggy areas around the drain field, unusually lush grass, strong odors, and contaminated well water.",
+	},
+	{
+		question: "What should I do if I have a water leak?",
+		answer:
+			"Shut off water at the nearest fixture valve or the main valve, keep electricity away from affected areas, and call us promptly for a professional repair.",
+	},
+	{
+		question: "How long does a water heater typically last?",
+		answer:
+			"Conventional tank water heaters commonly last 8 to 12 years, while tankless models can last 15 to 20 years or longer with proper maintenance.",
+	},
+	{
+		question: "What areas do you serve?",
+		answer:
+			"We serve Santa Cruz County and selected Santa Clara County communities in California. Call us if you are unsure whether a specific address is inside the current service area.",
+	},
+]
+
