+++ package.json
@@ -6,6 +6,7 @@
 		"dev": "next dev",
 		"prebuild": "node scripts/sync-maplibre-worker.mjs && node scripts/strip-next-polyfills.mjs",
 		"build": "next build",
+		"postbuild": "node scripts/strip-home-js.mjs",
 		"start": "next start",
 		"postinstall": "node scripts/sync-maplibre-worker.mjs",
 		"lint": "eslint . --max-warnings=0",
@@ -42,6 +43,7 @@
 		"@vercel/blob": "^2.6.1",
 		"class-variance-authority": "0.7.1",
 		"clsx": "2.1.1",
+		"critters": "0.0.23",
 		"evlog": "^2.22.4",
 		"gray-matter": "4.0.3",
 		"maplibre-gl": "^6.0.0",
